Setup Simple Menu UI
