+++ top-down-shooter-hdrp/Assets/Prefabs/Weapons/huge-physical-damage-data.asset
@@ -10,7 +10,7 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 51788645c410530428f999834b3b1fa9, type: 3}
-  m_Name: hyge-physical-damage-data
+  m_Name: huge-physical-damage-data
   m_EditorClassIdentifier: 
   damageType: 0
   damage: 20
