+++ top-down-shooter-hdrp/Assets/Prefabs/Weapons/huge-physical-damage-data.asset.meta
