+++ top-down-shooter-hdrp/Assets/Prefabs/Weapons/puncher-weapon-data.asset
@@ -9,16 +9,16 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: d3cb94c5fa723fa44a52c387a97eeb29, type: 3}
   m_Name: puncher-weapon-data
   m_EditorClassIdentifier: Assembly-CSharp::ShotgunWeaponData
   weaponPrefab: {fileID: 4436583768657473792, guid: 7497a60754d692e4095a794a3bdb8fee,
     type: 3}
   projectilePrefab: {fileID: 8735313740739425132, guid: decfdf5fbd0123b4c992f3b225625331,
     type: 3}
-  fireRate: 0.3
-  energyConsumption: 6
-  weaponIcon: {fileID: 21300000, guid: 095892cf3d608924d86ce9e84dc81813, type: 3}
+  fireRate: 0.4
+  energyConsumption: 8
+  weaponIcon: {fileID: 21300000, guid: 2875f7577c07b8c468e31621a0a21cd4, type: 3}
   fireDensity: 6
-  positionScatter: 0.2
+  positionScatter: 0.1
   rotationScatter: 20
