+++ top-down-shooter-hdrp/Assets/Scripts/Game Management/GameManager.cs
@@ -10,6 +10,7 @@ namespace top_down_shooter
         [SerializeField] private GameObject PlayerPrefab;               // instantiated in the bginning of the game
         [SerializeField] private Transform PlayerSpawn;                 // place where player should be instantiated
         [SerializeField] private GameObject PlayerUIControllerPrefab;   // prefab for UI controller 
+        [SerializeField] private GameObject EndgameUIMenuPrefab;        // prefab for engame screen
 
         private PlayerManager _playerManager;                           // for storing data about everything that should happen
         private PlayerUIManager _UIManager;                             // UI controller
@@ -27,8 +28,7 @@ namespace top_down_shooter
                 _playerManager = player.GetComponent<PlayerManager>();
                 _UIManager = UIManager.GetComponent<PlayerUIManager>();
 
-                _playerManager.Stats.OnPlayerDead +=
-                    () => SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex);
+                _playerManager.Stats.OnPlayerDead += _UIManager.InitializeEngameScreen;
 
                 _playerManager.Stats.OnEnergyChanged += _UIManager.UpdateEnergy;
                 _playerManager.Stats.OnHealthChanged += _UIManager.UpdateHealth;
