+++ top-down-shooter-hdrp/Assets/Scripts/Game Management/PlayerUIManager.cs
@@ -1,12 +1,15 @@
 ﻿using UnityEngine;
 using UnityEngine.UI;
 using TMPro;
+using UnityEngine.SceneManagement;
 
 class PlayerUIManager : MonoBehaviour
 {
     #region Fields
 
     // private fields
+    [SerializeField] private GameObject PlayerUICanvas;
+    [SerializeField] private GameObject EndgameUICanvas;
     [SerializeField] private Image _healthBar;
     [SerializeField] private TMP_Text _energyBar;
     [SerializeField] private Image _weaponIcon;
@@ -18,6 +21,8 @@ class PlayerUIManager : MonoBehaviour
 
     private void Awake()
     {
+        PlayerUICanvas.SetActive(true);
+        EndgameUICanvas.SetActive(false);
         _healthBar.type = Image.Type.Filled;
         _healthBar.fillMethod = Image.FillMethod.Horizontal;
     }
@@ -31,7 +36,6 @@ class PlayerUIManager : MonoBehaviour
         Debug.Log("Abilty Icon Changed");
         _abilityIcon.sprite = source;
     }
-
     public void UpdateHealth(float newHealthValue)
     {
         _healthBar.fillAmount = newHealthValue / 100f;
@@ -41,5 +45,13 @@ class PlayerUIManager : MonoBehaviour
         Debug.Log(newEnergyValue);
         _energyBar.text = newEnergyValue.ToString();
     }
+    public void InitializeEngameScreen()
+    {
+        PlayerUICanvas.SetActive(false);
+        EndgameUICanvas.SetActive(true);
+    }
+    public void ReloadScene () => SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex);
+    public void GoBackToMenu() => SceneManager.LoadScene(0);
+
     #endregion
 }
