+++ top-down-shooter-hdrp/Assets/Scripts/Weapon System/ShotgunWeaponData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d3cb94c5fa723fa44a52c387a97eeb29
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
